refactor(webview/core): regroup Core types & constants
