+++ src/constants.ts
@@ -1,10 +1,5 @@
 export const MAX_LOG_LIST_HISTORY = 1000
 
-export enum FileKind {
-  Directory = 'Directory',
-  File = 'File',
-}
-
 export enum Screen {
   Config = 'Config',
   Dashboard = 'Dashboard',
