+++ src/stories/screens/DashboardPage.stories.tsx
@@ -6,8 +6,8 @@ import { ArgStoreKey } from '../../../.storybook/argsStore/constants'
 import { useArgsStoreArgs } from '../../../.storybook/argsStore/useArgsStoreArgs'
 import { Layout } from '../../Layout'
 import { Screen } from '../../constants'
+import { Core } from '../../core/types'
 import { DashboardScreenComponent, type DashboardScreenComponentProps } from '../../screens/Dashboard/Component'
-import { Core } from '../../types'
 import { noop } from '../../utils/noop'
 import { waitFor } from '../../utils/waitFor'
